Changing how codes work + fixing inconsistencies
I realized that it makes sense for the code to have enough flexibility to record minutes, not just hours, so I added two digits after the initial one to represent them. In my editing of CompanyInfo, I also had it use the same code style as that code specifically will be referenced by other classes.